Port to MacOS.
Updated SDL headers to version 2.0.20 in line with the latest version on
Homebrew at the time.

To start the game on MacOS, it should be run on the terminal from the
executable's directory, otherwise relative paths cannot be resolved.
(Will fix in the future).

Some Mac devices don't support single-channel SRGB images so support
should be queried and images used appropriately.